sched: Align scheduling periods with timer rate
It is not possible to balance any sooner than the jiffy time in
milliseconds, as the scheduler requires a timer interrupt (tick) to
perform load balancing. We can make the scheduler as precise as possible
within the constraint of our timer rate to reduce as much latency as
possible.

By aligning the scheduling period, we load balance every tick. For a 100
Hz timer, we will be able to call the scheduler every 10 ms, and grant
1ms per task at least.

With a 1000 Hz timer, we will be balancing once every millisecond, which
may result in poor throughput, but the latency will be at most 1 ms.

THIS COMMIT IS EXPERIMENTAL. I highly suggest considering if your
device has the processing power to perform in latencies under your timer
rate in milliseconds.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Dušan Uverić <dusan.uveric9@gmail.com>

	modified:   kernel/sched/fair.c